server/requirements.txt
Flask
python-dotenv
requests
coolname
